# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/coins/actual": {
            "post": {
                "description": "Returns latest prices for requested coins",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coins"
                ],
                "summary": "Get latest coin prices",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"BTC,ETH\"",
                        "description": "Comma-separated list of coin titles",
                        "name": "titles",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Cryptoproject_pkg_dto.CoinResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Cryptoproject_pkg_dto.ErrorResponseDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Cryptoproject_pkg_dto.ErrorResponseDto"
                        }
                    }
                }
            }
        },
        "/api/v1/coins/aggregate/{aggFunc}": {
            "post": {
                "description": "Returns aggregated data (AVG/MAX/MIN) for requested coins",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coins"
                ],
                "summary": "Get aggregated coin data",
                "parameters": [
                    {
                        "enum": [
                            "AVG",
                            "MAX",
                            "MIN"
                        ],
                        "type": "string",
                        "description": "Aggregation function (AVG, MAX, MIN)",
                        "name": "aggFunc",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"BTC,ETH\"",
                        "description": "Comma-separated list of coin titles",
                        "name": "titles",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Cryptoproject_pkg_dto.AggregateCoinResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Cryptoproject_pkg_dto.ErrorResponseDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Cryptoproject_pkg_dto.ErrorResponseDto"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Cryptoproject_pkg_dto.AggregateCoinResponse": {
            "type": "object",
            "properties": {
                "coin_name": {
                    "type": "string"
                },
                "price": {
                    "description": "AVG, MAX или MIN",
                    "type": "number"
                }
            }
        },
        "Cryptoproject_pkg_dto.CoinResponse": {
            "type": "object",
            "properties": {
                "coin_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "Cryptoproject_pkg_dto.ErrorResponseDto": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        }
    }
}